Add support for RPM packages
This was thankfully made easy by NFPM!

Refactored the DEB packaging so that the code is almost entirely shared
between the two.
